pyproject.toml
[project]
name = "marl-cluster-training"
version = "0.1.0"
description = "MARL cluster training"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Yoann Poupart"},
]
requires-python = ">=3.11"
dependencies = [
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
]
scripts = [
    "hydra-core>=1.3.2",
    "loguru>=0.7.3",
    "benchmarl>=1.4.0",
    "id-marl-eval>=0.0.4",
    "matplotlib>=3.10.1",
    "pettingzoo[butterfly,sisl]>=1.24.3",
]

[build]
target-dir = "build/dist"

[tool.uv]
default-groups = ["dev", "scripts"]

[tool.uv.sources]
id-marl-eval = { git = "https://github.com/instadeepai/marl-eval.git" }

[tool.ruff]
line-length = 119
target-version = "py311"
